api/requirements.txt
Django>=3.2.4
PyMySQL>=1.0.2
PyJWT>=2.1.0
numpy>=1.21.0
django-redis>=5.0.0
Pillow>=8.3
beautifulsoup4>=4.9.3
requests>=2.25.1
channels>=3.0.4
pandas
djangorestframework
simplejson